chore: Gradient.js component is added & loadingUser is removed
